fix: changed int to size_t in loops.

+++ src/hash_set_coarse_grained.h
@@ -37,7 +37,7 @@ class HashSetCoarseGrained : public HashSetBase<T> {
 
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (int i = 0; i < bucket.size(); i++) {
+    for (size_t i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
         bucket.erase(i);
         return true;
